fix: satisfy backend-size and pkg-file-count for CLI sessionexecution tests
